DAP-Link/Firmware/src/requirements.txt
project_generator==0.9.13
mbed-ls==1.6.2
pyserial
pyOCD==0.16.0
requests
intelhex
six
enum34
pyusb
mbed-cli==1.8.2
bs4
numpy
pyelftools
fuzzywuzzy
jsonschema
